chore(tests): extract shared test helpers into single module
